Add assignment expression analysis
