+++ projects/ADbQuery/src/analyze_function.js
@@ -1,4 +1,5 @@
 import { typeExists } from "./analyzer_common.js";
+import { analyzeExpression } from "./analyze_expression.js";
 
 function functionName(node, object) {
     if (object) {
@@ -47,72 +48,6 @@ function validateReturn(node, object, ast) {
             )}' that has a return value.`;
         }
     }
-}
-
-function isArgument(type, node) {
-    return node["arguments"].includes(type);
-}
-
-function isField(type, context) {
-    return context["type"] == "object" && context["fields"].includes(type);
-}
-
-function isArrayType(type, context) {
-    return context["type"] == "array" && context["arrayType"] == type;
-}
-
-function isVariant(type, context) {
-    return context["type"] == "variant" && context["variants"].includes(type);
-}
-
-function expressionType(expression, node, context) {
-    if (context) {
-        if (isField(expression["value"], context)) {
-            return "field";
-        }
-
-        if (isArrayType(expression["value"], context)) {
-            return "arrayType";
-        }
-
-        if (isVariant(expression["value"], context)) {
-            return "variant";
-        }
-    }
-
-    if (!expression["parent"] && isArgument(expression["value"], node)) {
-        return "argument";
-    }
-
-    return "new";
-}
-
-function setExpressionType(expression, node, context) {
-    if (expression["type"] == "return") {
-        expression["returnType"] = expressionType(expression, node, context);
-    } else if (expression["type"] == "type") {
-        expression["type"] = expressionType(expression, node, context);
-    }
-}
-
-function expressionContext(expression, object, ast) {
-    const parent = expression["parent"];
-
-    if (parent) {
-        return ast[parent["value"]];
-    }
-
-    return object;
-}
-
-function analyzeExpression(expression, node, object, ast) {
-    const context = expressionContext(expression, object, ast);
-
-    if (expression["parent"]) {
-        analyzeExpression(expression["parent"], node, object, ast);
-    }
-
-    setExpressionType(expression, node, context);
 }
 
 function analyzeExpressions(node, object, ast) {
